refactor: move gemini openai and anthropic to provider folder
- Update import paths for `openai` and `gemini` packages
- Move `anthropic` package to `provider/anthropic`
- Move `gemini` package to `provider/gemini`
- Move `openai` package to `provider/openai`
