+++ provider/gemini/func.go
