+++ provider/openai/client.go
