+++ provider/openai/func.go
