+++ provider/openai/func_test.go
