+++ provider/openai/openai.go
