+++ provider/openai/options.go
