+++ provider/openai/options_test.go
